iOS-#56 [Measuring Tool] Refactor MeasuringToolViewController
